% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportNCCNMonthlyPeriodOfRecord}
\alias{exportNCCNMonthlyPeriodOfRecord}
\title{Export monthly period of record summaries for precip and air temp}
\usage{
exportNCCNMonthlyPeriodOfRecord(
  park_code = c("MORA", "NOCA", "OLYM", "SAJH"),
  file_out = paste(park_code, "monthly_periodofrecord.xlsx", sep = "_"),
  overwrite = FALSE
)
}
\arguments{
\item{park_code}{Four-letter park code(s)}

\item{file_out}{Path (including filename) of the .xlsx spreadsheet you want to create. Defaults to a file named with the park code and water year in the current working directory.}

\item{overwrite}{If \code{file_out} already exists, do you want to overwrite it? Defaults to \code{FALSE}.}
}
\value{
Writes data to \code{file_out} and returns a list containing the data that was written.
}
\description{
Export monthly period of record summaries for precip and air temp
}
\examples{
\dontrun{
exportNCCNMonthlyPeriodOfRecord("OLYM", 2022, "exports/OLYM_Daily_2022.xlsx")
}
}
