% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportNCCNMonthlySummaries}
\alias{exportNCCNMonthlySummaries}
\title{Export monthly precip and air temp summaries}
\usage{
exportNCCNMonthlySummaries(
  park_code = c("MORA", "NOCA", "OLYM", "SAJH"),
  water_year,
  file_out = paste0("./", park_code, "_", water_year, "_monthly.xlsx"),
  overwrite = FALSE
)
}
\arguments{
\item{park_code}{Four-letter park code(s)}

\item{water_year}{Integer vector indicating the water year(s) to fetch data for}

\item{file_out}{Path (including filename) of the .xlsx spreadsheet you want to create. Defaults to a file named with the park code and water year in the current working directory.}

\item{overwrite}{If \code{file_out} already exists, do you want to overwrite it? Defaults to \code{FALSE}.}
}
\value{
Writes data to \code{file_out} and returns a list containing the data that was written.
}
\description{
Export monthly precip and air temp summaries
}
\examples{
\dontrun{
exportNCCNDailySummaries("OLYM", 2022, "exports/OLYM_Daily_2022.xlsx")
}
}
