% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getClimateStationList}
\alias{getClimateStationList}
\title{Get information about NCCN climate stations}
\usage{
getClimateStationList(park_code = c("MORA", "NOCA", "OLYM", "SAJH"))
}
\arguments{
\item{park_code}{Four-letter park code(s)}
}
\value{
A tibble with one line per climate station
}
\description{
Retrieves information about climate stations for NCCN parks (MORA, NOCA, OLYM, and SAJH)
}
\examples{
\dontrun{
olym_station_info <- getClimateStationList("OLYM")
nccn_station_info <- getClimateStationList()
}
}
