% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius.R
\name{getClimateTimeSeriesInfo}
\alias{getClimateTimeSeriesInfo}
\title{Get NCCN climate time series information}
\usage{
getClimateTimeSeriesInfo(
  park_code = c("MORA", "NOCA", "OLYM", "SAJH"),
  include_derived = FALSE
)
}
\arguments{
\item{park_code}{Four-letter park code(s)}

\item{include_derived}{Include derived time series?}
}
\value{
A tibble with one line per dataset
}
\description{
Retrieves information about climate time series datasets for NCCN parks (MORA, NOCA, OLYM, and SAJH)
}
\examples{
\dontrun{
olym_ts_info <- getClimateTimeSeriesInfo("OLYM")
}
}
