% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getParkClimateData}
\alias{getParkClimateData}
\title{Get all corrected climate time series data for a park}
\usage{
getParkClimateData(
  park_code = c("MORA", "NOCA", "OLYM", "SAJH"),
  parameter,
  labels,
  water_year,
  include_derived = FALSE,
  time_zone = "America/Los_Angeles"
)
}
\arguments{
\item{park_code}{Four-letter park code(s)}

\item{parameter}{A character string indicating which parameter to retrieve (e.g. "Air Temp").}

\item{water_year}{Integer vector indicating the water year(s) to fetch data for}

\item{include_derived}{Include derived time series?}

\item{time_zone}{String indicating the time zone the data were collected in. See the full list of \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}{TZ identifiers} for more info.}
}
\value{
A list of corrected, unsummarized time series by park
}
\description{
Retrieves all corrected, unsummarized time series data from the [PARK]_Climate folder in Aquarius (e.g. OLYM_Climate).
}
\examples{
\dontrun{
olym <- getParkClimateData("OLYM")

}
}
