% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWaterYear}
\alias{getWaterYear}
\title{Get Water Year}
\usage{
getWaterYear(date)
}
\arguments{
\item{date}{A date-time object}
}
\value{
The water year of \code{date} as an integer number
}
\description{
Computes the water year that a date falls in, where the water year runs from
Oct 1 - Sept 30.
}
\examples{
my_date <- as.Date("2022-02-01")
getWaterYear(my_date)  # 2022
another_date <- as.Date("2022-10-01") # 2023
}
