% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarius.R
\name{summarizeParkClimateData}
\alias{summarizeParkClimateData}
\title{Summarize NCCN Climate Data}
\usage{
summarizeParkClimateData(
  park_code = c("MORA", "NOCA", "OLYM", "SAJH"),
  parameter,
  label,
  water_year,
  summary = c("mean", "min", "max", "sum"),
  summary_period = c("day", "month"),
  period_of_record = FALSE,
  time_zone = "America/Los_Angeles"
)
}
\arguments{
\item{park_code}{Four-letter park code(s)}

\item{parameter}{A character string indicating which parameter to retrieve (e.g. "Air Temp").}

\item{water_year}{Integer vector indicating the water year(s) to fetch data for}

\item{summary}{Type of summary to perform. Options are "mean", "min", "max", and "sum".}

\item{summary_period}{Period over which to summarize. Options are "day" and "month".}

\item{period_of_record}{Summarize over the whole period of record? If \code{TRUE}, summarizes across all years in dataset. If \code{FALSE}, summaries are grouped by year.}

\item{time_zone}{String indicating the time zone the data were collected in. See the full list of \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}{TZ identifiers} for more info.}
}
\value{
A data frame with a column of data for each station
}
\description{
Summarize NCCN Climate Data
}
\examples{
\dontrun{
# Summarize monthly total precip at OLYM over entire period of record
olym_precip <- summarizeParkClimateData("OLYM", "Precip Increm", "Rainfall", summary = "sum", summary_period = "month", period_of_record = TRUE)
olym_precip_2022 <- summarizeParkClimateData("OLYM", "Precip Increm", "Rainfall", 2022, summary = "sum", summary_period = "month", period_of_record = FALSE)
}
}
